* EOAccess/EOEntity.m 	Added comment. Please check this! 	* EOAccess/EODatabaseContext.m 	include NSObject+GNUstepBase.h 	avoid Enumerator exception. 	* EOAccess/EOUtilities.m 	split line into two 	* EOAccess/EOAdaptorChannel.h/m 	unsigned -> NSUInteger 	* EOAdaptors/PostgreSQLAdaptor/PostgreSQLChannel.h 	removed _describeResults 	* EOAdaptors/PostgreSQLAdaptor/PostgreSQLChannel.m 	include NSObject+GNUstepBase.h 	added numberOfAffectedRows 	attribute naming in raw rows fixed 	updating fixed 	unsigned -> NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@30586 72102866-910b-0410-8b05-ffd578937521

+++ EOAccess/EOAdaptorChannel.h
@@ -106,13 +106,16 @@
 - (void)updateValues: (NSDictionary *)values
 inRowDescribedByQualifier: (EOQualifier *)qualifier
 	      entity: (EOEntity *)entity;
-- (unsigned)updateValues: (NSDictionary *)values
-  inRowsDescribedByQualifier: (EOQualifier *)qualifier
-		  entity: (EOEntity *)entity;
+
+- (NSUInteger) updateValues: (NSDictionary *)values
+ inRowsDescribedByQualifier: (EOQualifier *)qualifier
+                     entity: (EOEntity *)entity;
+
 - (void)deleteRowDescribedByQualifier: (EOQualifier *)qualifier
-			       entity: (EOEntity *)entity;
-- (unsigned)deleteRowsDescribedByQualifier: (EOQualifier *)qualifier
-				    entity: (EOEntity *)entity;
+                               entity: (EOEntity *)entity;
+
+- (NSUInteger) deleteRowsDescribedByQualifier: (EOQualifier *)qualifier
+                                       entity: (EOEntity *)entity;
 
 /* Fetching rows */
 - (void)selectAttributes: (NSArray *)attributes
